requirements.txt
flask==3.0.0
python-dotenv==1.0.1
requests==2.31.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.10
alembic==1.13.1
